requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
coverage==7.4.3
python-dotenv==1.0.1
django-filter==23.5
django-rest-swagger==2.2.0
Pillow==10.2.0 